+++ lab-jeff/README.md
@@ -3,6 +3,7 @@
 #Express API#
 
 ###Summary###
+A basic singe resource http server utilizing express.
 
 
 ###Instructions###
@@ -19,11 +20,13 @@ To run the tests, linter, watcher and start the server type the command:
 
 POST: `curl -H "Content-Type: application/json" -X POST -d '{"pokeName":"charizard", "pokeType":"fire"}' http://localhost:3000/api/pokemon/`
 
-To retrieve a Pokemon follow the below format, "name" must match pokemon your created with POST.
+You will recieve a unique ID for that pokemon, save that number for future use of that pokemon.
+
+To retrieve or delete a Pokemon follow the below format, "pokeId" must match with the ID created with the Pokemon.
 
 GET: `curl http://localhost:3000/api/pokemon/pokeId/`
 
-DELETE: 'curl -X "DELETE" http:/localhost:3000/api/remove/pokemon/pokeID/'
+DELETE: 'curl -X "DELETE" http:/localhost:3000/api/pokemon/pokeID/'
 
 
 `Jeff Gebhardt - CF JS 401`
