+++ lab-jeff/model/appError.js
@@ -1,7 +1,9 @@
 'use strict';
 
 module.exports = exports = function AppError(res){
-  this.message = 'Error';
-  this.statusCode = res.status;
-  this.responseMessage = res.send('You have recieved an error with code: ' + this.statusCode);
+  this.message = console.log('Error...');
+  this.statusCode = 404;
+  this.responseMessage = function(){
+    res.status(404).send('You have recieved an error with code: ' + this.statusCode + ', page not found...');
+  };
 };
